Pulls all Views into their own files
