+++ Views/InstructionView.py
@@ -1,0 +1,21 @@
+from definitions import arcade, SCREEN_WIDTH, SCREEN_HEIGHT
+from Views.GameView import GameView
+
+'''
+Code taken from https://arcade.academy/examples/view_instructions_and_game_over.html#view-instructions-and-game-over
+'''
+class InstructionView(arcade.View):
+    def on_show(self):
+        arcade.set_background_color(arcade.color.ORANGE_PEEL)
+
+    def on_draw(self):
+        arcade.start_render()
+        arcade.draw_text("Instructions Screen", SCREEN_WIDTH/2, SCREEN_HEIGHT/2,
+                         arcade.color.BLACK, font_size=50, anchor_x="center")
+        arcade.draw_text("Click to advance", SCREEN_WIDTH/2, SCREEN_HEIGHT/2-75,
+                         arcade.color.GRAY, font_size=20, anchor_x="center")
+
+    def on_mouse_press(self, _x, _y, _button, _modifiers):
+        game_view = GameView()
+        game_view.setup()
+        self.window.show_view(game_view)
